+++ app/(stack)/record/index.tsx
