+++ app/(stack)/record/new/[id].tsx
