+++ app/_layout.tsx
@@ -1,42 +1,44 @@
 import "react-native-url-polyfill/auto";
 import { useState, useEffect } from "react";
-import { supabase } from "@/libs/supabase";
 import Auth from "@/components/Auth";
-import { View } from "react-native";
 import { Session } from "@supabase/supabase-js";
 import { Slot } from "expo-router";
-import Header from "@/components/Header";
+import { supabase } from "@/libs/supabase";
 import { MenusProvider } from "@/contexts/menusContext";
 import { ExercisesProvider } from "@/contexts/exercisesContext";
 import { RecordsProvider } from "@/contexts/recordsContext";
+import { Text, View } from "react-native";
 
 export default function App() {
   const [session, setSession] = useState<Session | null>(null);
+  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     supabase.auth.getSession().then(({ data: { session } }) => {
       setSession(session);
+      setLoading(false);
     });
 
     supabase.auth.onAuthStateChange((_event, session) => {
       setSession(session);
+      setLoading(false);
     });
   }, []);
+
+  // TODO:
+  if (loading) {
+    return (
+      <View>
+        <Text>Loading...</Text>
+      </View>
+    );
+  }
 
   return (
     <MenusProvider>
       <ExercisesProvider>
         <RecordsProvider>
-          <View>
-            {session && session.user ? (
-              <>
-                <Header />
-                <Slot />
-              </>
-            ) : (
-              <Auth />
-            )}
-          </View>
+          {session && session.user ? <Slot /> : <Auth />}
         </RecordsProvider>
       </ExercisesProvider>
     </MenusProvider>
